Fail deployment if version prefix already exists
While this is unlikely to occur, it is possible that in the future we
may hit a collision when using the short ref as the version component
in our S3 prefix. This change modifies the deployment to check if a
target version prefix exists, and if so, to fail without uploading.
That should trigger us to then increase the prefix size without any
unintentional surprises.