Remove orphan html file
This was previously extracted in the last import, but then written
over in mbz changes creating a new UUID. Since the content was
just filler and never used in production courses we'll just rip
it out.

Note: It's possible that in the future we have events that reference
this UUID in the one week it was live on dev courses, and we'll be
left wondering what happened. This commit may help solve such a
mystery if it occurs.