Implement work around for docker compose bug
Early versions of v2.24 do not work with how we have our .env
positioned alongside our docker-compose.yml. This bug is being
addressed upstream, but in the meantime we want to avoid impacting
authoring users. This change can likely be reverted in the future
once the upstream fix is released and included in latest versions of
Docker Desktop.